backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
aiosqlite>=0.20.0

python-jose[cryptography]>=3.3.0
cryptography>=42.0.0

passlib[bcrypt]==1.7.4
bcrypt==4.0.1

pydantic[email]>=2.0.0
slowapi>=0.1.9
python-multipart>=0.0.9
firebase-admin>=6.5.0
google-cloud-firestore>=2.16.0
httpx>=0.27.0
psutil==5.9.8
